+++ .gitignore
@@ -15,3 +15,5 @@ hs_err_pid*
 
 .classpath
 .project
+/target/
+.settings
