--- src/fi/ni/rdf/Namespace.java
@@ -1,14 +1,0 @@
-package fi.ni.rdf;
-
-public class Namespace {
-	public static final String RDF = "http://www.w3.org/1999/02/22-rdf-syntax-ns#";
-	public static final String RDFS = "http://www.w3.org/2000/01/rdf-schema#";
-	public static final String OWL = "http://www.w3.org/2002/07/owl#";
-	public static final String XSD = "http://www.w3.org/2001/XMLSchema#";
-	public static String IFC = "http://ifcowl.openbimstandards.org/";
-	public static final String DCE = "http://purl.org/dc/elements/1.1/";
-	public static final String VANN = "http://purl.org/vocab/vann/";
-	public static final String CC = "http://creativecommons.org/ns#";
-	public static final String LIST = "https://w3id.org/list#";	
-	public static final String EXPRESS = "https://w3id.org/express#";
-}
